+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 - `get` and `getUnsynced` methods will now return null instead of throwing an exception.
 - Added `getOrThrow` and `getUnsyncedOrThrow` methods to GreenhouseConfigHolder.
 - LateHolders no longer throw when unable to bind. Use the onException consumer to handle errors instead if necessary.
--
+
+## Bugfixes
+- Fixed potential race conditions on Fabric with platform helper.